fix pop-to-buffer in slime-help--open-buffer
before help buffer for a symbol already existed, it would call
pop-to-buffer. But when it was creating a new one it called
pop-to-buffer within slime-help--open-buffer which was called within
with-current-buffer, so it didn't have an effect. I move the calls to
slime-help--open-buffer outside of the with-current-buffer blocks so its
pop-to-buffer call can have an effect.